acpi: add acpi_of_match_device()
For ACPI enumerated OF-compatible devices we need to be able to look up our
of_device_id based on the "compatible" device property.  This often needs to be
done regardless of the OF-specific or bus-specific matching mechanism.

Therefore, introduce acpi_of_match_device(), which tells a driver which entry
in the of_match_table matches a device.